+++ lib/Subscription/SubscriptionHandler.php
@@ -19,6 +19,8 @@ use PagarMe\Sdk\Subscription\Request\SubscriptionTransactionsGet;
 
 class SubscriptionHandler extends AbstractHandler
 {
+    use \PagarMe\Sdk\Transaction\TransactionBuilder;
+
     /**
     * @param int $id
     * @param Plan $plan
@@ -141,30 +143,5 @@ class SubscriptionHandler extends AbstractHandler
         }
 
         return $transactions;
-    }
-
-    private function buildTransaction($transactionData)
-    {
-        if (isset($transactionData->split_rules)) {
-            $transactionData->split_rules = $this->buildSplitRules(
-                $transactionData->split_rules
-            );
-        }
-
-        if ($transactionData->payment_method == BoletoTransaction::PAYMENT_METHOD) {
-            return new BoletoTransaction(get_object_vars($transactionData));
-        }
-
-        if ($transactionData->payment_method == CreditCardTransaction::PAYMENT_METHOD) {
-            return new CreditCardTransaction(get_object_vars($transactionData));
-        }
-
-        throw new UnsupportedTransaction(
-            sprintf(
-                'Transaction type: %s, is not supported',
-                $transactionData->payment_method
-            ),
-            1
-        );
     }
 }
