+++ lib/Transaction/TransactionHandler.php
@@ -20,6 +20,8 @@ use PagarMe\Sdk\Recipient\Recipient;
 
 class TransactionHandler extends AbstractHandler
 {
+    use TransactionBuilder;
+
     public function creditCardTransaction(
         $amount,
         Card $card,
@@ -118,31 +120,6 @@ class TransactionHandler extends AbstractHandler
         $response = $this->client->send($request);
 
         return $this->buildTransaction($response);
-    }
-
-    private function buildTransaction($transactionData)
-    {
-        if (isset($transactionData->split_rules)) {
-            $transactionData->split_rules = $this->buildSplitRules(
-                $transactionData->split_rules
-            );
-        }
-
-        if ($transactionData->payment_method == BoletoTransaction::PAYMENT_METHOD) {
-            return new BoletoTransaction(get_object_vars($transactionData));
-        }
-
-        if ($transactionData->payment_method == CreditCardTransaction::PAYMENT_METHOD) {
-            return new CreditCardTransaction(get_object_vars($transactionData));
-        }
-
-        throw new UnsupportedTransaction(
-            sprintf(
-                'Transaction type: %s, is not supported',
-                $transactionData->payment_method
-            ),
-            1
-        );
     }
 
     private function buildSplitRules($splitRuleData)
